added inheritance
